Don't crash on unknown max resolution factor
For PDF JP2s, when max res factor may be only 4 or 5, we don't want the
server to puke - so we fall back to a lower level if the requested level
fails.  This is definitely not going to be awesome for performance, but
it should only affect a very specific subset of edge cases.  I *think*
it won't be a problem at all with our 512x512 tiles, and shouldn't be a
problem on thumbnails until/unless we start doing really small thumbs
for some special view or something.