Fix bug when cropping without resize
Defaulting resize to full area causes a problem when trying to crop but
not set explicit resize.  The default resize is still around, making the
decode logic believe it needs to upsize the image.

The IIIF server never sees this, as it always makes sure to set both
crop and resize, so the bug is likely only when using raw APIs.