rais-server: set sane constraint defaults
If an info.json would report a zero value for any dimension, we
explicitly override that with the maximum possible value.  Usually a
zero means the value wasn't set, but this also works even if zero is
explicitly set because it wouldn't make sense to actually serve images
if you wanted a limit of zero.