rais-server: encode image to buffer, not client
This fixes the problem where a bad decode spits out an http "multiple
response.WriteHeader calls" error by always verifying the image can be
decoded prior to sending it to the client.  This will reduce server
performance when clients are terminating their connection early, but it
is the correct approach.