plugins/json-tracer: improve reported path
RawPath sometimes holds the original path, such as
"test%2Ffile.jp2", whereas Path would hold "test/file.jp2".  RawPath is
empty when there isn't anything to decode, however, so we have to check
both....

Go docs are not clear on this fact ("The Parse function sets both Path
and RawPath in the URL it returns").  But the comments in net/url/url.go
are pretty clear, thankfully:

    // Set Path and, optionally, RawPath.
    // RawPath is a hint of the encoding of Path. We don't want to set it if
    // the default escaping of Path is equivalent, to help make sure that people
    // don't rely on it in general.