json-tracer: write traces without waiting for traffic
Instead of waiting for a hit to check if it's time to write the JSON out
to disk, we instead loop in the background for each tracer and write at
regular intervals.  Nothing is written when there are no traces to
write.  The time interval is now based on the first trace collection
rather than the start of the server.