rais-server: improve "not found" debugging
Adds a lot of middleware magic in order to get every request logged.  If
a request comes in that seems like it should work, you can now look at
the logs to see what RAIS actually got.  This is very handy when
proxying with Apache or nginx, where the real URL isn't always the same
as what's seen by the proxied application.