plugins/imagick-decoder: reduce "445" errors
Removes the reuse of ImageMagick data, which caused excessive problems
when writing out some kind of cache to /tmp.

I'm not entirely sure what ImageMagick was doing that became a problem,
but it seems you absolutely *must* free resources immediately after
using them.  Even though the second request instantiated a new C.Image,
it seems whatever ImageMagick does on the filesystem causes it to get
itself confused when the cache is still around and the same file is
requested for processing.

The bad news is that ImageMagick is clearly not threadsafe.  Even with
this fix, problems are less common but not eliminated altogether.  The
good news, however, is that this speeds up info.json requests
dramatically.  Because my first implementation IS TERRIBLE and I should
be very ashamed.  And I am.

Oddly, this only happens with large files, so I suspect ImageMagick has
some kind of limit to the number of bytes it's willing to write to its
filesystem-based cache.