Make all ImageMagick operations synchronous
This is really horrible, but at this point I've come to the conclusion
that the problem is concurrency, not limits.  I can verify that setting
the disk limit to 10 bytes prevents any large image request from
working, while setting it to infinity makes large images work just
fine... so long as they're requested sequentially.